Moved node modules to new package
