+++ src/main/java/com/caoccao/javet/interfaces/IJavetBiConsumer.java
