--- src/main/java/com/caoccao/javet/interop/node/NodeModuleAnnotation.java
@@ -1,9 +1,0 @@
-package com.caoccao.javet.interop.node;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface NodeModuleAnnotation {
-    String name() default "";
-}
