--- src/main/java/com/caoccao/javet/interop/node/NodeModuleAny.java
@@ -1,9 +1,0 @@
-package com.caoccao.javet.interop.node;
-
-import com.caoccao.javet.values.reference.V8ValueObject;
-
-public class NodeModuleAny extends BaseNodeModule {
-    public NodeModuleAny(V8ValueObject moduleObject, String name) {
-        super(moduleObject, name);
-    }
-}
