+++ src/main/java/com/caoccao/javet/node/modules/BaseNodeModule.java
@@ -15,30 +15,33 @@
  *
  */
 
-package com.caoccao.javet.values.reference.builtin;
+package com.caoccao.javet.node.modules;
 
 import com.caoccao.javet.exceptions.JavetException;
-import com.caoccao.javet.values.V8Value;
+import com.caoccao.javet.utils.JavetResourceUtils;
 import com.caoccao.javet.values.reference.V8ValueObject;
 
-import java.util.Objects;
+public abstract class BaseNodeModule implements INodeModule {
+    protected V8ValueObject moduleObject;
+    protected String name;
 
-@SuppressWarnings("unchecked")
-public class V8ValueBuiltInJson extends V8ValueObject {
-
-    public static final String FUNCTION_STRINGIFY = "stringify";
-
-    public V8ValueBuiltInJson(long handle) {
-        super(handle);
-    }
-
-    public String stringify(V8Value v8Value) throws JavetException {
-        Objects.requireNonNull(v8Value);
-        return invokeString(FUNCTION_STRINGIFY, v8Value);
+    public BaseNodeModule(V8ValueObject moduleObject, String name) {
+        this.moduleObject = moduleObject;
+        this.name = name;
     }
 
     @Override
-    public V8ValueBuiltInJson toClone() throws JavetException {
-        return this;
+    public void close() throws JavetException {
+        JavetResourceUtils.safeClose(moduleObject);
     }
+
+    @Override
+    public V8ValueObject getModuleObject() {
+        return moduleObject;
+    }
+
+    public String getName() {
+        return name;
+    }
+
 }
