+++ src/main/java/com/caoccao/javet/node/modules/INodeModule.java
@@ -15,7 +15,13 @@
  *
  */
 
-package com.caoccao.javet.values;
+package com.caoccao.javet.node.modules;
 
-public abstract class V8Data {
+import com.caoccao.javet.interfaces.IJavetClosable;
+import com.caoccao.javet.values.reference.V8ValueObject;
+
+public interface INodeModule extends IJavetClosable {
+    V8ValueObject getModuleObject();
+
+    String getName();
 }
