+++ src/main/java/com/caoccao/javet/node/modules/NodeModuleAny.java
@@ -15,7 +15,12 @@
  *
  */
 
-package com.caoccao.javet.values;
+package com.caoccao.javet.node.modules;
 
-public abstract class V8Data {
+import com.caoccao.javet.values.reference.V8ValueObject;
+
+public class NodeModuleAny extends BaseNodeModule {
+    public NodeModuleAny(V8ValueObject moduleObject, String name) {
+        super(moduleObject, name);
+    }
 }
