+++ src/main/java/com/caoccao/javet/utils/JavetDefaultLogger.java
